+++ governance-app/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Support for update validator score parameters transaction.
+
 ## 1.0.0
 
 ### Added